Upgrade chart
Made some necessary changes to the Chart so I could deploy to my
instance of Kubernetes (v1.23.6-2+2a84a218e3cd52). Some Kubernetes
resources changed since the time this Chart was last updated.

The docker image that is deployed using this Chart has also changed
since the last time this Chart was last updated. I needed to update the
names of the environment variables in the ConfigMap and Deployment
resources. The template for the Ingress controller also needed to be
updated.